Pick,fit(4F)/Back-end/model/catvton_FASTAPI/requirements.txt
torch==2.1.2
torchvision==0.16.2
accelerate==0.31.0
diffusers==0.29.2
huggingface_hub==0.23.4
#matplotlib==3.9.1
numpy==1.26.4
opencv_python==4.10.0.84
pillow==10.3.0
PyYAML==6.0.1
scipy==1.13.1
setuptools==51.0.0
scikit-image==0.24.0
tqdm==4.66.4
transformers==4.27.3
fvcore==0.1.5.post20221221
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
av==12.3.0
gradio==4.41.0